backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
langchain-openai
requests
passlib[bcrypt]
python-jose
python-dotenv
